BRProject/requirements.txt
# Core ML stack
torch==2.6.0
numpy==1.26.4
ultralytics==8.3.40
opencv-python==4.10.0.84
pillow==11.0.0

# API stack
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1

# Required by FastAPI for multipart/form-data file uploads
python-multipart==0.0.17

# Optional: only needed when /analyze?enhance_with_ai=true is used
openai==1.55.3
